requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
coqui-tts==0.24.1
coqui-tts-trainer==0.1.4
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
decorator==4.4.2
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
grpcio==1.65.5
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
huggingface-hub==0.24.5
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
inflect==7.3.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
openai-whisper==20231117
packaging==24.1
pillow==10.2.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
protobuf==5.27.3
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
pysbd==0.3.4
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-vlc==3.0.20123
pytube==15.0.0
pytubefix==6.12.0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentencepiece==0.2.0
setuptools==72.2.0
shellingham==1.5.4
shiboken6==6.7.2
six==1.16.0
smart-open==7.0.4
soundfile==0.12.1
soxr==0.4.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.1+cu124
torchaudio==2.4.1
torchvision==0.19.1+cu124
tqdm==4.66.5
transformers==4.40.2
typeguard==4.3.0
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
yarl==1.9.4
